Fetch inbox content and implement interactions
Before we could only show dummy messages in the inbox. Now we fetch the actual content associated the notifications in the users inbox.

Iframes can now request the client to send an interaction. The client will then post the the iconet/public/interaction.php endpoint. There the interaction is forwarded to the processor.